bots/ZeroG/requirements.txt
discord.py
requests
aiohttp

# Optional but recommended:
python-dotenv
